enhance: support create vol with some cache args
