+++ master/api_args_parse.go
@@ -1,0 +1,972 @@
+package master
+
+import (
+	"bytes"
+	"encoding/json"
+	"errors"
+	"fmt"
+	"io/ioutil"
+	"net/http"
+	"strconv"
+	"strings"
+	"time"
+
+	"github.com/chubaofs/chubaofs/proto"
+	"github.com/chubaofs/chubaofs/util/cryptoutil"
+	"github.com/chubaofs/chubaofs/util/log"
+)
+
+// Parse the request that adds/deletes a raft node.
+func parseRequestForRaftNode(r *http.Request) (id uint64, host string, err error) {
+	if err = r.ParseForm(); err != nil {
+		return
+	}
+	var idStr string
+	if idStr = r.FormValue(idKey); idStr == "" {
+		err = keyNotFound(idKey)
+		return
+	}
+
+	if id, err = strconv.ParseUint(idStr, 10, 64); err != nil {
+		return
+	}
+	if host = r.FormValue(addrKey); host == "" {
+		err = keyNotFound(addrKey)
+		return
+	}
+
+	if arr := strings.Split(host, colonSplit); len(arr) < 2 {
+		err = unmatchedKey(addrKey)
+		return
+	}
+	return
+}
+
+func parseRequestForUpdateMetaNode(r *http.Request) (nodeAddr string, id uint64, err error) {
+	if err = r.ParseForm(); err != nil {
+		return
+	}
+	if nodeAddr, err = extractNodeAddr(r); err != nil {
+		return
+	}
+	if id, err = extractNodeID(r); err != nil {
+		return
+	}
+	return
+}
+
+func parseRequestForAddNode(r *http.Request) (nodeAddr, zoneName string, err error) {
+	if err = r.ParseForm(); err != nil {
+		return
+	}
+	if nodeAddr, err = extractNodeAddr(r); err != nil {
+		return
+	}
+	if zoneName = r.FormValue(zoneNameKey); zoneName == "" {
+		zoneName = DefaultZoneName
+	}
+	return
+}
+
+func parseAndExtractNodeAddr(r *http.Request) (nodeAddr string, err error) {
+	if err = r.ParseForm(); err != nil {
+		return
+	}
+	return extractNodeAddr(r)
+}
+
+func parseRequestToDecommissionNode(r *http.Request) (nodeAddr, diskPath string, err error) {
+	if err = r.ParseForm(); err != nil {
+		return
+	}
+	nodeAddr, err = extractNodeAddr(r)
+	if err != nil {
+		return
+	}
+	diskPath, err = extractDiskPath(r)
+	return
+}
+
+func parseRequestToGetTaskResponse(r *http.Request) (tr *proto.AdminTask, err error) {
+	var body []byte
+	if err = r.ParseForm(); err != nil {
+		return
+	}
+	if body, err = ioutil.ReadAll(r.Body); err != nil {
+		return
+	}
+	tr = &proto.AdminTask{}
+	decoder := json.NewDecoder(bytes.NewBuffer([]byte(body)))
+	decoder.UseNumber()
+	err = decoder.Decode(tr)
+	return
+}
+
+func parseVolName(r *http.Request) (name string, err error) {
+	if err = r.ParseForm(); err != nil {
+		return
+	}
+	if name, err = extractName(r); err != nil {
+		return
+	}
+	return
+}
+
+func parseGetVolParameter(r *http.Request) (p *getVolParameter, err error) {
+	p = &getVolParameter{}
+	skipOwnerValidationVal := r.Header.Get(proto.SkipOwnerValidation)
+	if len(skipOwnerValidationVal) > 0 {
+		if p.skipOwnerValidation, err = strconv.ParseBool(skipOwnerValidationVal); err != nil {
+			return
+		}
+	}
+	if p.name = r.FormValue(nameKey); p.name == "" {
+		err = keyNotFound(nameKey)
+		return
+	}
+	if !volNameRegexp.MatchString(p.name) {
+		err = errors.New("name can only be number and letters")
+		return
+	}
+	if p.authKey = r.FormValue(volAuthKey); !p.skipOwnerValidation && len(p.authKey) == 0 {
+		err = keyNotFound(volAuthKey)
+		return
+	}
+	return
+}
+
+func parseVolNameAndAuthKey(r *http.Request) (name, authKey string, err error) {
+	if err = r.ParseForm(); err != nil {
+		return
+	}
+	if name, err = extractName(r); err != nil {
+		return
+	}
+	if authKey, err = extractAuthKey(r); err != nil {
+		return
+	}
+	return
+
+}
+
+func parseRequestToDeleteVol(r *http.Request) (name, authKey string, err error) {
+	return parseVolNameAndAuthKey(r)
+
+}
+
+func parseRequestToUpdateVol(r *http.Request) (name, authKey, description string, err error) {
+	if err = r.ParseForm(); err != nil {
+		return
+	}
+	if name, err = extractName(r); err != nil {
+		return
+	}
+	if authKey, err = extractAuthKey(r); err != nil {
+		return
+	}
+	description = r.FormValue(descriptionKey)
+	return
+}
+
+func parseDefaultInfoToUpdateVol(r *http.Request, vol *Vol) (zoneName string, capacity uint64, replicaNum int,
+	dpSelectorName string, dpSelectorParm string, err error) {
+	if err = r.ParseForm(); err != nil {
+		return
+	}
+	if zoneName = r.FormValue(zoneNameKey); zoneName == "" {
+		zoneName = vol.zoneName
+	}
+	if capacityStr := r.FormValue(volCapacityKey); capacityStr != "" {
+		var capacityInt int
+		if capacityInt, err = strconv.Atoi(capacityStr); err != nil {
+			err = unmatchedKey(volCapacityKey)
+			return
+		}
+		capacity = uint64(capacityInt)
+	} else {
+		capacity = vol.Capacity
+	}
+	if replicaNumStr := r.FormValue(replicaNumKey); replicaNumStr != "" {
+		if replicaNum, err = strconv.Atoi(replicaNumStr); err != nil {
+			err = unmatchedKey(replicaNumKey)
+			return
+		}
+	} else {
+		replicaNum = int(vol.dpReplicaNum)
+	}
+	dpSelectorName = r.FormValue(dpSelectorNameKey)
+	dpSelectorParm = r.FormValue(dpSelectorParmKey)
+	if (dpSelectorName == "") || (dpSelectorParm == "") {
+		if (dpSelectorName != "") || (dpSelectorParm != "") {
+			err = keyNotFound(dpSelectorNameKey + " or " + dpSelectorParmKey)
+			return
+		}
+		dpSelectorName = vol.dpSelectorName
+		dpSelectorParm = vol.dpSelectorParm
+	}
+	return
+}
+
+func parseBoolFieldToUpdateVol(r *http.Request, vol *Vol) (followerRead, authenticate bool, err error) {
+	if followerReadStr := r.FormValue(followerReadKey); followerReadStr != "" {
+		if followerRead, err = strconv.ParseBool(followerReadStr); err != nil {
+			err = unmatchedKey(followerReadKey)
+			return
+		}
+	} else {
+		followerRead = vol.FollowerRead
+	}
+	if authenticateStr := r.FormValue(authenticateKey); authenticateStr != "" {
+		if authenticate, err = strconv.ParseBool(authenticateStr); err != nil {
+			err = unmatchedKey(authenticateKey)
+			return
+		}
+	} else {
+		authenticate = vol.authenticate
+	}
+	return
+}
+
+func parseRequestToSetVolCapacity(r *http.Request) (name, authKey string, capacity int, err error) {
+	if err = r.ParseForm(); err != nil {
+		return
+	}
+	if name, err = extractName(r); err != nil {
+		return
+	}
+	if authKey, err = extractAuthKey(r); err != nil {
+		return
+	}
+	if capacity, err = extractCapacity(r); err != nil {
+		return
+	}
+	return
+}
+
+type coldVolArgs struct {
+	objBlockSize     int
+	ebsCapacity      int
+	cacheAction      int
+	cacheThreshold   int
+	cacheTtl         int
+	cacheHighWater   float64
+	cacheLowWater    float64
+	cacheLRUInterval int
+}
+
+type createVolReq struct {
+	name            string
+	owner           string
+	size            int
+	mpCount         int
+	dpReplicaNum    int
+	capacity        int
+	vol             *Vol
+	followerRead    bool
+	authenticate    bool
+	crossZone       bool
+	defaultPriority bool
+	zoneName        string
+	description     string
+	volType         int
+
+	// cold vol args
+	coldArgs coldVolArgs
+}
+
+func checkCacheAction(action int) error {
+
+	if action != noCache && action != rCache && action != rwCache {
+		return fmt.Errorf("cache action is not legal, action [%d]", action)
+	}
+
+	return nil
+}
+
+func parseColdArgs(r *http.Request) (args coldVolArgs, err error) {
+
+	if args.objBlockSize, err = extractUint(r, ebsBlkSizeKey); err != nil {
+		return
+	}
+
+	if args.ebsCapacity, err = extractUint(r, ebsCapacityKey); err != nil {
+		return
+	}
+
+	if args.cacheAction, err = extractUint(r, cacheActionKey); err != nil {
+		return
+	}
+
+	if args.cacheThreshold, err = extractUint(r, cacheThresholdKey); err != nil {
+		return
+	}
+
+	if args.cacheTtl, err = extractUint(r, cacheTTLKey); err != nil {
+		return
+	}
+
+	if args.cacheHighWater, err = extractFloat64(r, cacheHighWaterKey); err != nil {
+		return
+	}
+
+	if args.cacheLowWater, err = extractFloat64(r, cacheLowWaterKey); err != nil {
+		return
+	}
+
+	if args.cacheTtl, err = extractUint(r, cacheLRUIntervalKey); err != nil {
+		return
+	}
+
+	return
+}
+
+func parseRequestToCreateVol(r *http.Request, req *createVolReq) (err error) {
+
+	if err = r.ParseForm(); err != nil {
+		return
+	}
+
+	if req.name, err = extractName(r); err != nil {
+		return
+	}
+
+	if req.owner, err = extractOwner(r); err != nil {
+		return
+	}
+
+	if req.coldArgs, err = parseColdArgs(r); err != nil {
+		return
+	}
+
+	if mpCountStr := r.FormValue(metaPartitionCountKey); mpCountStr != "" {
+		if req.mpCount, err = strconv.Atoi(mpCountStr); err != nil {
+			req.mpCount = defaultInitMetaPartitionCount
+		}
+	}
+
+	if replicaStr := r.FormValue(replicaNumKey); replicaStr == "" {
+		req.dpReplicaNum = defaultReplicaNum
+	} else if req.dpReplicaNum, err = strconv.Atoi(replicaStr); err != nil {
+		err = unmatchedKey(replicaNumKey)
+		return
+	}
+
+	if sizeStr := r.FormValue(dataPartitionSizeKey); sizeStr != "" {
+		if req.size, err = strconv.Atoi(sizeStr); err != nil {
+			err = unmatchedKey(dataPartitionSizeKey)
+			return
+		}
+	}
+
+	if req.capacity, err = extractCapacity(r); err != nil {
+		return
+	}
+
+	if req.volType, err = extractVolType(r); err != nil {
+		return
+	}
+
+	if req.followerRead, err = extractFollowerRead(r); err != nil {
+		return
+	}
+
+	if req.authenticate, err = extractAuthenticate(r); err != nil {
+		return
+	}
+
+	if req.crossZone, err = extractCrossZone(r); err != nil {
+		return
+	}
+
+	if req.defaultPriority, err = extractDefaulPriority(r); err != nil {
+		return
+	}
+
+	req.zoneName = r.FormValue(zoneNameKey)
+	req.description = r.FormValue(descriptionKey)
+	return
+}
+
+func parseRequestToCreateDataPartition(r *http.Request) (count int, name string, err error) {
+	if err = r.ParseForm(); err != nil {
+		return
+	}
+	if countStr := r.FormValue(countKey); countStr == "" {
+		err = keyNotFound(countKey)
+		return
+	} else if count, err = strconv.Atoi(countStr); err != nil || count == 0 || count > 32 {
+		err = unmatchedKey(countKey)
+		return
+	}
+	if name, err = extractName(r); err != nil {
+		return
+	}
+	return
+}
+
+func parseRequestToGetDataPartition(r *http.Request) (ID uint64, volName string, err error) {
+	if err = r.ParseForm(); err != nil {
+		return
+	}
+	if ID, err = extractDataPartitionID(r); err != nil {
+		return
+	}
+	volName = r.FormValue(nameKey)
+	return
+}
+
+func parseRequestToLoadDataPartition(r *http.Request) (ID uint64, err error) {
+	if err = r.ParseForm(); err != nil {
+		return
+	}
+	if ID, err = extractDataPartitionID(r); err != nil {
+		return
+	}
+	return
+}
+
+func parseRequestToAddMetaReplica(r *http.Request) (ID uint64, addr string, err error) {
+	return extractMetaPartitionIDAndAddr(r)
+}
+
+func parseRequestToRemoveMetaReplica(r *http.Request) (ID uint64, addr string, err error) {
+	return extractMetaPartitionIDAndAddr(r)
+}
+
+func extractMetaPartitionIDAndAddr(r *http.Request) (ID uint64, addr string, err error) {
+	if err = r.ParseForm(); err != nil {
+		return
+	}
+	if ID, err = extractMetaPartitionID(r); err != nil {
+		return
+	}
+	if addr, err = extractNodeAddr(r); err != nil {
+		return
+	}
+	return
+}
+
+func parseRequestToAddDataReplica(r *http.Request) (ID uint64, addr string, err error) {
+	return extractDataPartitionIDAndAddr(r)
+}
+
+func parseRequestToRemoveDataReplica(r *http.Request) (ID uint64, addr string, err error) {
+	return extractDataPartitionIDAndAddr(r)
+}
+
+func extractDataPartitionIDAndAddr(r *http.Request) (ID uint64, addr string, err error) {
+	if err = r.ParseForm(); err != nil {
+		return
+	}
+	if ID, err = extractDataPartitionID(r); err != nil {
+		return
+	}
+	if addr, err = extractNodeAddr(r); err != nil {
+		return
+	}
+	return
+}
+
+func extractDataPartitionID(r *http.Request) (ID uint64, err error) {
+	var value string
+	if value = r.FormValue(idKey); value == "" {
+		err = keyNotFound(idKey)
+		return
+	}
+	return strconv.ParseUint(value, 10, 64)
+}
+
+func parseRequestToDecommissionDataPartition(r *http.Request) (ID uint64, nodeAddr string, err error) {
+	return extractDataPartitionIDAndAddr(r)
+}
+
+func extractNodeAddr(r *http.Request) (nodeAddr string, err error) {
+	if nodeAddr = r.FormValue(addrKey); nodeAddr == "" {
+		err = keyNotFound(addrKey)
+		return
+	}
+	return
+}
+
+func extractNodeID(r *http.Request) (ID uint64, err error) {
+	var value string
+	if value = r.FormValue(idKey); value == "" {
+		err = keyNotFound(idKey)
+		return
+	}
+	return strconv.ParseUint(value, 10, 64)
+}
+
+func extractDiskPath(r *http.Request) (diskPath string, err error) {
+	if diskPath = r.FormValue(diskPathKey); diskPath == "" {
+		err = keyNotFound(diskPathKey)
+		return
+	}
+	return
+}
+
+func parseRequestToLoadMetaPartition(r *http.Request) (partitionID uint64, err error) {
+	if err = r.ParseForm(); err != nil {
+		return
+	}
+	if partitionID, err = extractMetaPartitionID(r); err != nil {
+		return
+	}
+	return
+}
+
+func parseRequestToDecommissionMetaPartition(r *http.Request) (partitionID uint64, nodeAddr string, err error) {
+	return extractMetaPartitionIDAndAddr(r)
+}
+
+func parseAndExtractStatus(r *http.Request) (status bool, err error) {
+
+	if err = r.ParseForm(); err != nil {
+		return
+	}
+	return extractStatus(r)
+}
+
+func extractStatus(r *http.Request) (status bool, err error) {
+	var value string
+	if value = r.FormValue(enableKey); value == "" {
+		err = keyNotFound(enableKey)
+		return
+	}
+	if status, err = strconv.ParseBool(value); err != nil {
+		return
+	}
+	return
+}
+
+func extractFollowerRead(r *http.Request) (followerRead bool, err error) {
+	var value string
+	if value = r.FormValue(followerReadKey); value == "" {
+		followerRead = false
+		return
+	}
+	if followerRead, err = strconv.ParseBool(value); err != nil {
+		return
+	}
+	return
+}
+
+func extractAuthenticate(r *http.Request) (authenticate bool, err error) {
+	var value string
+	if value = r.FormValue(authenticateKey); value == "" {
+		authenticate = false
+		return
+	}
+	if authenticate, err = strconv.ParseBool(value); err != nil {
+		return
+	}
+	return
+}
+
+func extractCrossZone(r *http.Request) (crossZone bool, err error) {
+	var value string
+	if value = r.FormValue(crossZoneKey); value == "" {
+		crossZone = false
+		return
+	}
+	if crossZone, err = strconv.ParseBool(value); err != nil {
+		return
+	}
+	return
+}
+
+func extractDefaulPriority(r *http.Request) (defaultPrior bool, err error) {
+	var value string
+	if value = r.FormValue(defaultPriority); value == "" {
+		defaultPrior = false
+		return
+	}
+	if defaultPrior, err = strconv.ParseBool(value); err != nil {
+		return
+	}
+	return
+}
+
+func parseAndExtractThreshold(r *http.Request) (threshold float64, err error) {
+	if err = r.ParseForm(); err != nil {
+		return
+	}
+	var value string
+	if value = r.FormValue(thresholdKey); value == "" {
+		err = keyNotFound(thresholdKey)
+		return
+	}
+	if threshold, err = strconv.ParseFloat(value, 64); err != nil {
+		return
+	}
+	return
+}
+func parseAndExtractSetNodeSetInfoParams(r *http.Request) (params map[string]interface{}, err error) {
+	if err = r.ParseForm(); err != nil {
+		return
+	}
+	var value string
+	params = make(map[string]interface{})
+	if value = r.FormValue(countKey); value != "" {
+		var count = uint64(0)
+		count, err = strconv.ParseUint(value, 10, 64)
+		if err != nil {
+			err = unmatchedKey(countKey)
+			return
+		}
+		params[countKey] = count
+	}
+	var zoneName string
+	if zoneName = r.FormValue(zoneNameKey); zoneName == "" {
+		zoneName = DefaultZoneName
+	}
+	if value = r.FormValue(idKey); value != "" {
+		var nodesetId = uint64(0)
+		nodesetId, err = strconv.ParseUint(value, 10, 64)
+		if err != nil {
+			err = unmatchedKey(idKey)
+			return
+		}
+		params[idKey] = nodesetId
+	}
+	return
+}
+func parseAndExtractSetNodeInfoParams(r *http.Request) (params map[string]interface{}, err error) {
+	if err = r.ParseForm(); err != nil {
+		return
+	}
+	var value string
+	noParams := true
+	params = make(map[string]interface{})
+	if value = r.FormValue(nodeDeleteBatchCountKey); value != "" {
+		noParams = false
+		var batchCount = uint64(0)
+		batchCount, err = strconv.ParseUint(value, 10, 64)
+		if err != nil {
+			err = unmatchedKey(nodeDeleteBatchCountKey)
+			return
+		}
+		params[nodeDeleteBatchCountKey] = batchCount
+	}
+
+	if value = r.FormValue(nodeMarkDeleteRateKey); value != "" {
+		noParams = false
+		var val = uint64(0)
+		val, err = strconv.ParseUint(value, 10, 64)
+		if err != nil {
+			err = unmatchedKey(nodeMarkDeleteRateKey)
+			return
+		}
+		params[nodeMarkDeleteRateKey] = val
+	}
+
+	if value = r.FormValue(nodeAutoRepairRateKey); value != "" {
+		noParams = false
+		var val = uint64(0)
+		val, err = strconv.ParseUint(value, 10, 64)
+		if err != nil {
+			err = unmatchedKey(nodeAutoRepairRateKey)
+			return
+		}
+		params[nodeAutoRepairRateKey] = val
+	}
+
+	if value = r.FormValue(nodeDeleteWorkerSleepMs); value != "" {
+		noParams = false
+		var val = uint64(0)
+		val, err = strconv.ParseUint(value, 10, 64)
+		if err != nil {
+			err = unmatchedKey(nodeMarkDeleteRateKey)
+			return
+		}
+		params[nodeDeleteWorkerSleepMs] = val
+	}
+
+	if value = r.FormValue(clusterLoadFactor); value != "" {
+		noParams = false
+		valF, err := strconv.ParseFloat(value, 64)
+		if err != nil || valF < 1 {
+			err = unmatchedKey(clusterLoadFactor)
+			return params, err
+		}
+
+		params[clusterLoadFactor] = valF
+	}
+
+	if noParams {
+		err = keyNotFound(nodeDeleteBatchCountKey)
+		return
+	}
+	return
+}
+
+func validateRequestToCreateMetaPartition(r *http.Request) (volName string, start uint64, err error) {
+	if volName, err = extractName(r); err != nil {
+		return
+	}
+
+	var value string
+	if value = r.FormValue(startKey); value == "" {
+		err = keyNotFound(startKey)
+		return
+	}
+	start, err = strconv.ParseUint(value, 10, 64)
+	return
+}
+
+func parseAndExtractPartitionInfo(r *http.Request) (partitionID uint64, err error) {
+	if err = r.ParseForm(); err != nil {
+		return
+	}
+	if partitionID, err = extractMetaPartitionID(r); err != nil {
+		return
+	}
+	return
+}
+
+func extractMetaPartitionID(r *http.Request) (partitionID uint64, err error) {
+	var value string
+	if value = r.FormValue(idKey); value == "" {
+		err = keyNotFound(idKey)
+		return
+	}
+	return strconv.ParseUint(value, 10, 64)
+}
+
+func extractCapacity(r *http.Request) (capacity int, err error) {
+	var capacityStr string
+	if capacityStr = r.FormValue(volCapacityKey); capacityStr == "" {
+		err = keyNotFound(volCapacityKey)
+		return
+	}
+	if capacity, err = strconv.Atoi(capacityStr); err != nil {
+		err = unmatchedKey(volCapacityKey)
+	}
+	return
+}
+
+func extractVolType(r *http.Request) (volType int, err error) {
+
+	var capacityStr string
+	if capacityStr = r.FormValue(volTypeKey); capacityStr == "" {
+		err = keyNotFound(volTypeKey)
+		return
+	}
+
+	if volType, err = strconv.Atoi(volTypeKey); err != nil {
+		err = unmatchedKey(volCapacityKey)
+	}
+
+	return
+}
+
+func extractAuthKey(r *http.Request) (authKey string, err error) {
+	if authKey = r.FormValue(volAuthKey); authKey == "" {
+		err = keyNotFound(volAuthKey)
+		return
+	}
+	return
+}
+
+func parseAndExtractName(r *http.Request) (name string, err error) {
+	if err = r.ParseForm(); err != nil {
+		return
+	}
+	return extractName(r)
+}
+
+func extractName(r *http.Request) (name string, err error) {
+	if name = r.FormValue(nameKey); name == "" {
+		err = keyNotFound(nameKey)
+		return
+	}
+	if !volNameRegexp.MatchString(name) {
+		return "", errors.New("name can only be number and letters")
+	}
+
+	return
+}
+
+func extractUint(r *http.Request, key string) (val int, err error) {
+	var str string
+	if str = r.FormValue(key); str == "" {
+		return 0, nil
+	}
+
+	if val, err = strconv.Atoi(str); err != nil || val < 0 {
+		return 0, fmt.Errorf("args [%s] is not legal, val %s", key, str)
+	}
+
+	return val, nil
+}
+
+func extractStr(r *http.Request, key string) (val string) {
+
+	return r.FormValue(key)
+}
+
+func extractFloat64(r *http.Request, key string) (val float64, err error) {
+	var str string
+	if str = r.FormValue(key); str == "" {
+		return 0, nil
+	}
+
+	if val, err = strconv.ParseFloat(str, 64); err != nil || val < 0 {
+		return 0, fmt.Errorf("args [%s] is not legal, val %s", key, str)
+	}
+
+	return val, nil
+}
+
+func extractOwner(r *http.Request) (owner string, err error) {
+	if owner = r.FormValue(volOwnerKey); owner == "" {
+		err = keyNotFound(volOwnerKey)
+		return
+	}
+	if !ownerRegexp.MatchString(owner) {
+		return "", errors.New("owner can only be number and letters")
+	}
+
+	return
+}
+
+func parseAndCheckTicket(r *http.Request, key []byte, volName string) (jobj proto.APIAccessReq, ticket cryptoutil.Ticket, ts int64, err error) {
+	var (
+		plaintext []byte
+	)
+
+	if err = r.ParseForm(); err != nil {
+		return
+	}
+
+	if plaintext, err = extractClientReqInfo(r); err != nil {
+		return
+	}
+
+	if err = json.Unmarshal([]byte(plaintext), &jobj); err != nil {
+		return
+	}
+
+	if err = proto.VerifyAPIAccessReqIDs(&jobj); err != nil {
+		return
+	}
+
+	ticket, ts, err = extractTicketMess(&jobj, key, volName)
+
+	return
+}
+
+func extractClientReqInfo(r *http.Request) (plaintext []byte, err error) {
+	var (
+		message string
+	)
+	if err = r.ParseForm(); err != nil {
+		return
+	}
+
+	if message = r.FormValue(proto.ClientMessage); message == "" {
+		err = keyNotFound(proto.ClientMessage)
+		return
+	}
+
+	if plaintext, err = cryptoutil.Base64Decode(message); err != nil {
+		return
+	}
+
+	return
+}
+
+func extractTicketMess(req *proto.APIAccessReq, key []byte, volName string) (ticket cryptoutil.Ticket, ts int64, err error) {
+	if ticket, err = proto.ExtractTicket(req.Ticket, key); err != nil {
+		err = fmt.Errorf("extractTicket failed: %s", err.Error())
+		return
+	}
+	if time.Now().Unix() >= ticket.Exp {
+		err = proto.ErrExpiredTicket
+		return
+	}
+	if ts, err = proto.ParseVerifier(req.Verifier, ticket.SessionKey.Key); err != nil {
+		err = fmt.Errorf("parseVerifier failed: %s", err.Error())
+		return
+	}
+	if err = proto.CheckAPIAccessCaps(&ticket, proto.APIRsc, req.Type, proto.APIAccess); err != nil {
+		err = fmt.Errorf("CheckAPIAccessCaps failed: %s", err.Error())
+		return
+	}
+	if err = proto.CheckVOLAccessCaps(&ticket, volName, proto.VOLAccess, proto.MasterNode); err != nil {
+		err = fmt.Errorf("CheckVOLAccessCaps failed: %s", err.Error())
+		return
+	}
+	return
+}
+
+func newSuccessHTTPReply(data interface{}) *proto.HTTPReply {
+	return &proto.HTTPReply{Code: proto.ErrCodeSuccess, Msg: proto.ErrSuc.Error(), Data: data}
+}
+
+func newErrHTTPReply(err error) *proto.HTTPReply {
+	if err == nil {
+		return newSuccessHTTPReply("")
+	}
+	code, ok := proto.Err2CodeMap[err]
+	if ok {
+		return &proto.HTTPReply{Code: code, Msg: err.Error()}
+	}
+	return &proto.HTTPReply{Code: proto.ErrCodeInternalError, Msg: err.Error()}
+}
+
+func sendOkReply(w http.ResponseWriter, r *http.Request, httpReply *proto.HTTPReply) (err error) {
+	switch httpReply.Data.(type) {
+	case *DataPartition:
+		dp := httpReply.Data.(*DataPartition)
+		dp.RLock()
+		defer dp.RUnlock()
+	case *MetaPartition:
+		mp := httpReply.Data.(*MetaPartition)
+		mp.RLock()
+		defer mp.RUnlock()
+	case *MetaNode:
+		mn := httpReply.Data.(*MetaNode)
+		mn.RLock()
+		defer mn.RUnlock()
+	case *DataNode:
+		dn := httpReply.Data.(*DataNode)
+		dn.RLock()
+		defer dn.RUnlock()
+	}
+	reply, err := json.Marshal(httpReply)
+	if err != nil {
+		log.LogErrorf("fail to marshal http reply[%v]. URL[%v],remoteAddr[%v] err:[%v]", httpReply, r.URL, r.RemoteAddr, err)
+		http.Error(w, "fail to marshal http reply", http.StatusBadRequest)
+		return
+	}
+	send(w, r, reply)
+	return
+}
+
+func send(w http.ResponseWriter, r *http.Request, reply []byte) {
+	w.Header().Set("content-type", "application/json")
+	w.Header().Set("Content-Length", strconv.Itoa(len(reply)))
+	if _, err := w.Write(reply); err != nil {
+		log.LogErrorf("fail to write http reply[%s] len[%d].URL[%v],remoteAddr[%v] err:[%v]", string(reply), len(reply), r.URL, r.RemoteAddr, err)
+		return
+	}
+	log.LogInfof("URL[%v],remoteAddr[%v],response ok", r.URL, r.RemoteAddr)
+	return
+}
+
+func sendErrReply(w http.ResponseWriter, r *http.Request, httpReply *proto.HTTPReply) {
+	log.LogInfof("URL[%v],remoteAddr[%v],response err[%v]", r.URL, r.RemoteAddr, httpReply)
+	reply, err := json.Marshal(httpReply)
+	if err != nil {
+		log.LogErrorf("fail to marshal http reply[%v]. URL[%v],remoteAddr[%v] err:[%v]", httpReply, r.URL, r.RemoteAddr, err)
+		http.Error(w, "fail to marshal http reply", http.StatusBadRequest)
+		return
+	}
+	w.Header().Set("content-type", "application/json")
+	w.Header().Set("Content-Length", strconv.Itoa(len(reply)))
+	if _, err = w.Write(reply); err != nil {
+		log.LogErrorf("fail to write http reply[%s] len[%d].URL[%v],remoteAddr[%v] err:[%v]", string(reply), len(reply), r.URL, r.RemoteAddr, err)
+	}
+	return
+}
