+++ master/gapi_volume.go
@@ -190,9 +190,22 @@ func (s *VolumeService) createVolume(ctx context.Context, args struct {
 		return nil, fmt.Errorf("[%s] not has permission to create volume for [%s]", uid, args.Owner)
 	}
 
-	vol, err := s.cluster.createVol(args.Name, args.Owner, args.ZoneName, args.Description, int(args.MpCount),
-		int(args.DpReplicaNum), int(args.DataPartitionSize), int(args.Capacity),
-		args.FollowerRead, args.Authenticate, args.CrossZone, args.DefaultPriority)
+	req := &createVolReq{
+		name:            args.Name,
+		owner:           args.Owner,
+		zoneName:        args.ZoneName,
+		description:     args.Description,
+		mpCount:         int(args.MpCount),
+		dpReplicaNum:    int(args.DpReplicaNum),
+		size:            int(args.DataPartitionSize),
+		capacity:        int(args.Capacity),
+		followerRead:    args.FollowerRead,
+		authenticate:    args.Authenticate,
+		crossZone:       args.CrossZone,
+		defaultPriority: args.DefaultPriority,
+	}
+
+	vol, err := s.cluster.createVol(req)
 	if err != nil {
 		return nil, err
 	}
